@@ -1,23 +1,23 @@
 {
     "type": "project",
-    "license": "proprietary",
+    "license": "MIT",
     "require": {
-        "php": "^7.1.3",
+        "php": "^7.4.6",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "api-platform/api-pack": "^1.2",
+        "api-platform/api-pack": "*",
         "doctrine/annotations": "^1.10",
         "lexik/jwt-authentication-bundle": "^2.6",
-        "symfony/console": "4.4.*",
-        "symfony/dotenv": "4.4.*",
+        "symfony/console": "5.*",
+        "symfony/dotenv": "5.*",
         "symfony/flex": "^1.3.1",
-        "symfony/form": "4.4.*",
-        "symfony/framework-bundle": "4.4.*",
+        "symfony/form": "5.*",
+        "symfony/framework-bundle": "5.*",
         "symfony/maker-bundle": "^1.18",
-        "symfony/security": "4.4.*",
-        "symfony/validator": "4.4.*",
-        "symfony/web-server-bundle": "4.4.*",
-        "symfony/yaml": "4.4.*"
+        "symfony/security": "*",
+        "symfony/validator": "5.*",
+        "symfony/web-server-bundle": "*",
+        "symfony/yaml": "5.*"
     },
     "require-dev": {
     },
@@ -63,7 +63,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "4.4.*"
+            "require": "5.*"
         }
     }
 }
