@@ -10,16 +10,16 @@
         }
     ],
     "require": {
-        "php":                 "^7.3.0",
-        "composer-plugin-api": "^1.0.0"
+        "php":                 "^7.4.0",
+        "composer-plugin-api": "^1.1.0 || ^2.0"
     },
     "require-dev": {
-        "phpunit/phpunit":          "^8.2.5",
-        "infection/infection":      "^0.13.4",
-        "composer/composer":        "^1.8.6",
-        "ext-zip":                  "*",
-        "doctrine/coding-standard": "^6.0.0",
-        "vimeo/psalm":              "^3.4.9"
+        "phpunit/phpunit":          "^9.0.1",
+        "infection/infection":      "^0.15.3",
+        "composer/composer":        "^1.9.3 || ^2.0@dev",
+        "ext-zip":                  "^1.15.0",
+        "doctrine/coding-standard": "^7.0.2",
+        "vimeo/psalm":              "^3.9.3"
     },
     "autoload": {
         "psr-4": {
@@ -34,7 +34,7 @@
     "extra": {
         "class": "PackageVersions\\Installer",
         "branch-alias": {
-            "dev-master": "1.6.x-dev"
+            "dev-master": "1.99.x-dev"
         }
     },
     "scripts": {
