@@ -16,14 +16,15 @@
         }
     ],
     "require": {
-        "php": "^7.1.3"
+        "php": ">=7.2.5"
     },
     "suggest": {
         "symfony/http-foundation": ""
     },
     "require-dev": {
-        "symfony/http-foundation": "^3.4|^4.0|^5.0",
-        "symfony/http-kernel": "^3.4|^4.0|^5.0"
+        "symfony/http-client": "^4.4|^5.0",
+        "symfony/http-foundation": "^4.4|^5.0",
+        "symfony/http-kernel": "^4.4|^5.0"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Asset\\": "" },
@@ -34,7 +35,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "4.4-dev"
+            "dev-master": "5.1-dev"
         }
     }
 }
