@@ -21,26 +21,27 @@
         "symfony/cache-implementation": "1.0"
     },
     "require": {
-        "php": "^7.1.3",
+        "php": ">=7.2.5",
         "psr/cache": "~1.0",
         "psr/log": "~1.0",
         "symfony/cache-contracts": "^1.1.7|^2",
+        "symfony/polyfill-php80": "^1.15",
         "symfony/service-contracts": "^1.1|^2",
-        "symfony/var-exporter": "^4.2|^5.0"
+        "symfony/var-exporter": "^4.4|^5.0"
     },
     "require-dev": {
         "cache/integration-tests": "dev-master",
-        "doctrine/cache": "~1.6",
-        "doctrine/dbal": "~2.5",
-        "predis/predis": "~1.1",
+        "doctrine/cache": "^1.6",
+        "doctrine/dbal": "^2.5|^3.0",
+        "predis/predis": "^1.1",
         "psr/simple-cache": "^1.0",
-        "symfony/config": "^4.2|^5.0",
-        "symfony/dependency-injection": "^3.4|^4.1|^5.0",
+        "symfony/config": "^4.4|^5.0",
+        "symfony/dependency-injection": "^4.4|^5.0",
         "symfony/var-dumper": "^4.4|^5.0"
     },
     "conflict": {
         "doctrine/dbal": "<2.5",
-        "symfony/dependency-injection": "<3.4",
+        "symfony/dependency-injection": "<4.4",
         "symfony/http-kernel": "<4.4",
         "symfony/var-dumper": "<4.4"
     },
@@ -53,7 +54,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "4.4-dev"
+            "dev-master": "5.1-dev"
         }
     }
 }
