@@ -16,19 +16,21 @@
         }
     ],
     "require": {
-        "php": "^7.1.3",
-        "symfony/filesystem": "^3.4|^4.0|^5.0",
-        "symfony/polyfill-ctype": "~1.8"
+        "php": ">=7.2.5",
+        "symfony/deprecation-contracts": "^2.1",
+        "symfony/filesystem": "^4.4|^5.0",
+        "symfony/polyfill-ctype": "~1.8",
+        "symfony/polyfill-php80": "^1.15"
     },
     "require-dev": {
-        "symfony/event-dispatcher": "^3.4|^4.0|^5.0",
-        "symfony/finder": "^3.4|^4.0|^5.0",
-        "symfony/messenger": "^4.1|^5.0",
+        "symfony/event-dispatcher": "^4.4|^5.0",
+        "symfony/finder": "^4.4|^5.0",
+        "symfony/messenger": "^4.4|^5.0",
         "symfony/service-contracts": "^1.1|^2",
-        "symfony/yaml": "^3.4|^4.0|^5.0"
+        "symfony/yaml": "^4.4|^5.0"
     },
     "conflict": {
-        "symfony/finder": "<3.4"
+        "symfony/finder": "<4.4"
     },
     "suggest": {
         "symfony/yaml": "To use the yaml reference dumper"
@@ -42,7 +44,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "4.4-dev"
+            "dev-master": "5.1-dev"
         }
     }
 }
