@@ -16,18 +16,20 @@
         }
     ],
     "require": {
-        "php": "^7.1.3",
+        "php": ">=7.2.5",
         "symfony/polyfill-mbstring": "~1.0",
         "symfony/polyfill-php73": "^1.8",
-        "symfony/service-contracts": "^1.1|^2"
+        "symfony/polyfill-php80": "^1.15",
+        "symfony/service-contracts": "^1.1|^2",
+        "symfony/string": "^5.1"
     },
     "require-dev": {
-        "symfony/config": "^3.4|^4.0|^5.0",
-        "symfony/event-dispatcher": "^4.3",
-        "symfony/dependency-injection": "^3.4|^4.0|^5.0",
+        "symfony/config": "^4.4|^5.0",
+        "symfony/event-dispatcher": "^4.4|^5.0",
+        "symfony/dependency-injection": "^4.4|^5.0",
         "symfony/lock": "^4.4|^5.0",
-        "symfony/process": "^3.4|^4.0|^5.0",
-        "symfony/var-dumper": "^4.3|^5.0",
+        "symfony/process": "^4.4|^5.0",
+        "symfony/var-dumper": "^4.4|^5.0",
         "psr/log": "~1.0"
     },
     "provide": {
@@ -40,10 +42,11 @@
         "psr/log": "For using the console logger"
     },
     "conflict": {
-        "symfony/dependency-injection": "<3.4",
-        "symfony/event-dispatcher": "<4.3|>=5",
+        "symfony/dependency-injection": "<4.4",
+        "symfony/dotenv": "<5.1",
+        "symfony/event-dispatcher": "<4.4",
         "symfony/lock": "<4.4",
-        "symfony/process": "<3.3"
+        "symfony/process": "<4.4"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Console\\": "" },
@@ -54,7 +57,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "4.4-dev"
+            "dev-master": "5.1-dev"
         }
     }
 }
