@@ -16,14 +16,16 @@
         }
     ],
     "require": {
-        "php": "^7.1.3",
+        "php": ">=7.2.5",
         "psr/container": "^1.0",
+        "symfony/deprecation-contracts": "^2.1",
+        "symfony/polyfill-php80": "^1.15",
         "symfony/service-contracts": "^1.1.6|^2"
     },
     "require-dev": {
-        "symfony/yaml": "^3.4|^4.0|^5.0",
-        "symfony/config": "^4.3",
-        "symfony/expression-language": "^3.4|^4.0|^5.0"
+        "symfony/yaml": "^4.4|^5.0",
+        "symfony/config": "^5.1",
+        "symfony/expression-language": "^4.4|^5.0"
     },
     "suggest": {
         "symfony/yaml": "",
@@ -33,10 +35,10 @@
         "symfony/proxy-manager-bridge": "Generate service proxies to lazy load them"
     },
     "conflict": {
-        "symfony/config": "<4.3|>=5.0",
-        "symfony/finder": "<3.4",
-        "symfony/proxy-manager-bridge": "<3.4",
-        "symfony/yaml": "<3.4"
+        "symfony/config": "<5.1",
+        "symfony/finder": "<4.4",
+        "symfony/proxy-manager-bridge": "<4.4",
+        "symfony/yaml": "<4.4"
     },
     "provide": {
         "psr/container-implementation": "1.0",
@@ -51,7 +53,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "4.4-dev"
+            "dev-master": "5.1-dev"
         }
     }
 }
