@@ -16,10 +16,11 @@
         }
     ],
     "require": {
-        "php": "^7.1.3"
+        "php": ">=7.2.5",
+        "symfony/deprecation-contracts": "^2.1"
     },
     "require-dev": {
-        "symfony/process": "^3.4.2|^4.0|^5.0"
+        "symfony/process": "^4.4|^5.0"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Dotenv\\": "" },
@@ -30,7 +31,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "4.4-dev"
+            "dev-master": "5.1-dev"
         }
     }
 }
