@@ -16,14 +16,15 @@
         }
     ],
     "require": {
-        "php": "^7.1.3",
-        "psr/log": "~1.0",
-        "symfony/debug": "^4.4.5",
+        "php": ">=7.2.5",
+        "psr/log": "^1.0",
+        "symfony/polyfill-php80": "^1.15",
         "symfony/var-dumper": "^4.4|^5.0"
     },
     "require-dev": {
         "symfony/http-kernel": "^4.4|^5.0",
-        "symfony/serializer": "^4.4|^5.0"
+        "symfony/serializer": "^4.4|^5.0",
+        "symfony/deprecation-contracts": "^2.1"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\ErrorHandler\\": "" },
@@ -34,7 +35,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "4.4-dev"
+            "dev-master": "5.1-dev"
         }
     }
 }
